Moved createWebDriver inside SafeDriver
